Adding types file for Sidebar TS
