+++ src/components/sidebar/index.tsx
@@ -2,43 +2,9 @@ import { useMemo, useState } from 'react'
 import useCurrentPath from 'hooks/use-current-path'
 import SidebarBackToLink from './components/sidebar-back-to-link'
 import SidebarFilterInput from './components/sidebar-filter-input'
+import { MenuItem, SidebarProps } from './types'
 import SidebarNav from './components/sidebar-nav'
 import s from './style.module.css'
-
-/**
- *
- * For reference: this is also defined in react-components/docs-sidenav:
- * https://github.com/hashicorp/react-components/blob/main/packages/docs-sidenav/types.ts
- *
- */
-export interface MenuItem {
-  divider?: boolean
-  fullPath?: string
-  hasActiveChild?: boolean
-  hasChildrenMatchingFilter?: boolean
-  href?: string
-  id?: string
-  isActive?: boolean
-  matchesFilter?: boolean
-  path?: string
-  routes?: MenuItem[]
-  title?: string
-  /* Temporary solution to allow rendering of unlinked headings, as in designs */
-  heading?: string
-}
-
-export interface SidebarProps {
-  /** Optional { text, url } to use for the "← Back to..." link at the top of the sidebar */
-  backToLink?: {
-    text: string
-    url: string
-  }
-  menuItems: MenuItem[]
-  /** Whether or not the Sidebar should render the filter text input */
-  showFilterInput?: boolean
-  /** title to be shown as the title of the sidebar */
-  title: string
-}
 
 const addItemMetadata = (
   currentPath: string,
@@ -147,4 +113,5 @@ const Sidebar: React.FC<SidebarProps> = ({
   )
 }
 
+export type { MenuItem, SidebarProps }
 export default Sidebar
